Controllers now support multiple methods
Previously each controller only had a single exec() method. Now, the
routes file can specify a method to be called using @ as a separator.
