+++ config/routes.php
@@ -88,15 +88,15 @@ return [
   [ 'POST','/unsubscribe-thread', 'UnsubscribeThreadController', 'unsubscribeThread' ],
   [ 'GET','/tutorials/[:slug]/lessons/sort', 'LessonsSortController', 'lessonsSort' ],
   [ 'POST','/tutorials/[:slug]/lessons/save-sort-order', 'LessonsOrderController', 'lessonsOrder' ],
-  [ 'GET','/blog', 'BlogIndexController', 'blogIndex' ],
-  [ 'GET','/blog/new', 'BlogNewController', 'blogNew' ],
-  [ 'POST','/blog/create', 'BlogCreateController', 'blogCreate' ],
-  [ 'GET','/blog/[:slug]', 'BlogShowController', 'blogShow' ],
-  [ 'POST','/toggle-blog-alert', 'BlogToggleAlertController', 'blogToggleAlert' ],
-  [ 'POST','/blog/comment', 'BlogCommentController', 'blogComment' ],
-  [ 'GET','/blog/[:slug]/edit', 'BlogEditController', 'blogEdit' ],
-  [ 'POST','/blog/[:slug]/update', 'BlogUpdateController', 'blogUpdate' ],
-  [ 'POST','/blog/comment/[:comment_id]/unpublish', 'BlogCommentUnpublishController', 'blogCommentsUnpublish' ],
+  [ 'GET','/blog', 'BlogController@index', 'blogIndex' ],
+  [ 'GET','/blog/new', 'BlogController@new', 'blogNew' ],
+  [ 'POST','/blog/create', 'BlogController@create', 'blogCreate' ],
+  [ 'GET','/blog/[:slug]', 'BlogController@show', 'blogShow' ],
+  [ 'GET','/blog/[:slug]/edit', 'BlogController@edit', 'blogEdit' ],
+  [ 'POST','/blog/[:slug]/update', 'BlogController@update', 'blogUpdate' ],
+  [ 'POST','/toggle-blog-alert', 'BlogController@toggleAlert', 'blogToggleAlert' ],
+  [ 'POST','/blog/comment', 'BlogController@addComment', 'blogAddComment' ],
+  [ 'POST','/blog/comment/[:comment_id]/unpublish', 'BlogController@unpublishComment', 'blogUnpublishComment' ],
   [ 'GET','/favourites', 'FavouritesController', 'favourites' ],
   [ 'GET','/reference', 'ReferenceController', 'reference' ],
   [ 'GET','/asset-library', 'AssetLibraryController', 'assetLibrary' ],
