--- tests/src/PyAngelo/Controllers/Blog/BlogCommentUnpublishControllerTest.php
@@ -1,80 +1,0 @@
-<?php
-namespace Tests\src\PyAngelo\Controllers\Blog;
-
-use PHPUnit\Framework\TestCase;
-use Mockery;
-use Framework\Request;
-use Framework\Response;
-use PyAngelo\Controllers\Blog\BlogCommentUnpublishController;
-
-class BlogCommentUnpublishControllerTest extends TestCase {
-  public function setUp(): void {
-    $this->request = new Request($GLOBALS);
-    $this->response = new Response('views');
-    $this->auth = Mockery::mock('PyAngelo\Auth\Auth');
-    $this->blogRepository = Mockery::mock('PyAngelo\Repositories\BlogRepository');
-    $this->controller = new BlogCommentUnpublishController (
-      $this->request,
-      $this->response,
-      $this->auth,
-      $this->blogRepository
-    );
-  }
-  public function tearDown(): void {
-    Mockery::close();
-  }
-
-  public function testClassCanBeInstantiated() {
-    $this->assertSame(get_class($this->controller), 'PyAngelo\Controllers\Blog\BlogCommentUnpublishController');
-  }
-
-  public function testWhenNotAdmin() {
-    $this->auth->shouldReceive('isAdmin')->once()->with()->andReturn(false);
-
-    $response = $this->controller->exec();
-    $responseVars = $response->getVars();
-    $expectedHeaders = array(array('header', 'Location: /'));
-    $expectedFlashMessage = "You are not authorised!";
-    $this->assertSame($expectedHeaders, $response->getHeaders());
-    $this->assertSame($expectedFlashMessage, $_SESSION['flash']['message']);
-  }
-
-  public function testInvalidCrsfToken() {
-    $this->auth->shouldReceive('isAdmin')->once()->with()->andReturn(true);
-    $this->auth->shouldReceive('crsfTokenIsValid')->once()->with()->andReturn(false);
-
-    $response = $this->controller->exec();
-    $responseVars = $response->getVars();
-    $expectedHeaders = array(array('header', 'Location: /'));
-    $expectedFlashMessage = "You must delete comments from the PyAngelo website!";
-    $this->assertSame($expectedHeaders, $response->getHeaders());
-    $this->assertSame($expectedFlashMessage, $_SESSION['flash']['message']);
-  }
-
-  public function testNoCommentId() {
-    $this->auth->shouldReceive('isAdmin')->once()->with()->andReturn(true);
-    $this->auth->shouldReceive('crsfTokenIsValid')->once()->with()->andReturn(true);
-
-    $response = $this->controller->exec();
-    $responseVars = $response->getVars();
-    $expectedHeaders = array(array('header', 'Location: /page-not-found'));
-    $this->assertSame($expectedHeaders, $response->getHeaders());
-  }
-
-  public function testUnpublishComment() {
-    $commentId = 1;
-    $this->auth->shouldReceive('isAdmin')->once()->with()->andReturn(true);
-    $this->auth->shouldReceive('crsfTokenIsValid')->once()->with()->andReturn(true);
-    $this->blogRepository->shouldReceive('unpublishCommentById')
-      ->once()
-      ->with($commentId)
-      ->andReturn(1);
-    $this->request->post['comment_id'] = $commentId;
-
-    $response = $this->controller->exec();
-    $responseVars = $response->getVars();
-    $expectedHeaders = array(array('header', 'Location: /'));
-    $this->assertSame($expectedHeaders, $response->getHeaders());
-  }
-}
-?>
