--- tests/src/PyAngelo/Controllers/Blog/BlogNewControllerTest.php
@@ -1,60 +1,0 @@
-<?php
-namespace Tests\src\PyAngelo\Controllers\Blog;
-
-use PHPUnit\Framework\TestCase;
-use Mockery;
-use Framework\Request;
-use Framework\Response;
-use PyAngelo\Controllers\Blog\BlogNewController;
-
-class BlogNewControllerTest extends TestCase {
-  public function setUp(): void {
-    $this->request = new Request($GLOBALS);
-    $this->response = new Response('views');
-    $this->auth = Mockery::mock('PyAngelo\Auth\Auth');
-    $this->blogRepository = Mockery::mock('PyAngelo\Repositories\BlogRepository');
-    $this->controller = new BlogNewController (
-      $this->request,
-      $this->response,
-      $this->auth,
-      $this->blogRepository
-    );
-  }
-  public function tearDown(): void {
-    Mockery::close();
-  }
-
-  public function testClassCanBeInstantiated() {
-    $this->assertSame(get_class($this->controller), 'PyAngelo\Controllers\Blog\BlogNewController');
-  }
-
-  public function testBlogNewControllerWhenNotAdmin() {
-    $this->auth->shouldReceive('isAdmin')->once()->with()->andReturn(false);
-
-    $response = $this->controller->exec();
-    $responseVars = $response->getVars();
-    $expectedHeaders = array(array('header', 'Location: /'));
-    $expectedFlashMessage = "You are not authorised!";
-    $this->assertSame($expectedHeaders, $response->getHeaders());
-    $this->assertSame($expectedFlashMessage, $_SESSION['flash']['message']);
-  }
-
-  public function testBlogNewControllerWhenAdmin() {
-    $this->auth->shouldReceive('isAdmin')->once()->with()->andReturn(true);
-    $this->auth->shouldReceive('getPersonDetailsForViews')->once()->with();
-    $this->blogRepository->shouldReceive('getAllBlogCategories')
-      ->once()
-      ->with()
-      ->andReturn('blogCategories');
-
-    $response = $this->controller->exec();
-    $responseVars = $response->getVars();
-    $expectedViewName = 'blog/new.html.php';
-    $expectedPageTitle = 'Create a New Blog';
-    $expectedMetaDescription = "Create an amazing new blog for the PyAngelo crowd.";
-    $this->assertSame($expectedViewName, $response->getView());
-    $this->assertSame($expectedPageTitle, $responseVars['pageTitle']);
-    $this->assertSame($expectedMetaDescription, $responseVars['metaDescription']);
-  }
-}
-?>
